project-folder-v2/requirements.txt
fastapi
python-decouple
user-agents
pyjwt
uvicorn
jose
SQLAlchemy<1.5
psycopg2-binary==2.9.7